add google trace
